Split network package into client and server
