+++ core/src/main/java/com/elite/network/StartNetwork.java
@@ -1,5 +1,7 @@
 package com.elite.network;
 
+import com.elite.network.server.Server;
+
 /**
  * Starts the server
  *
