+++ core/src/main/java/com/elite/network/client/Client.java
@@ -1,10 +1,11 @@
-package com.elite.network;
+package com.elite.network.client;
 
 import com.elite.entities.pickups.Crate;
 import com.elite.entities.characters.Zombie;
 import com.elite.game.ClientGame;
 import com.elite.game.GameObject;
 import com.elite.game.NetworkGame;
+import com.elite.network.Keys;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
