+++ core/src/main/java/com/elite/network/server/ServerThread.java
@@ -1,4 +1,4 @@
-package com.elite.network;
+package com.elite.network.server;
 
 import com.elite.game.GameObject;
 import com.elite.game.NetworkGame;
