+++ core/src/main/java/com/elite/ui/HealthBar.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.elite.entities.ZombieAttributes;
+import com.elite.entities.characters.ZombieAttributes;
 
 
 /**
