+++ core/src/main/java/com/elite/ui/HomeScreen.java
@@ -14,7 +14,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 
 import com.elite.audio.AudioAccessor;
-import com.elite.network.Client;
+import com.elite.network.client.Client;
 import com.elite.world.SpriteRenderer;
 
 /**
